Add Resolver.
This commit introduces the idea of Resolvers. These are used to
fetch credentials from an external store. By default, every catalog
uses the PlaintextResolver, which is also the only one that ships with this commit.

Resolvers are extensible and can also be added for other credential stores like AWS SSM.